# internal/api/modelapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lineage Model API",
    "version": "v1",
    "description": "Machine-facing Model API for the Lineage AI model registry: publish, resolve, fetch, and manage models, versions, artifacts, lineage, and deployments. Auth is handled by infrastructure; a trusted X-Lineage-Actor header is recorded for audit."
  },
  "servers": [
    {
      "url": "/",
      "description": "Model API port (:8081)"
    }
  ],
...
  "paths": {
...
    "/v1/models/{model}/classifications": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        }
      ],
      "get": {
        "operationId": "listClassifications",
        "tags": [
          "compliance"
        ],
        "summary": "Every regime's classification for a model",
        "description": "Each row carries its own state, computed against its own classifiedAt.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationList"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/models/{model}/classifications/{regime}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        },
        {
          "name": "regime",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Regime"
          },
          "description": "The regime being read or written. Putting it in the path is what makes the write safe: a client cannot see or touch another regime's row."
        }
      ],
      "get": {
        "operationId": "getClassification",
        "tags": [
          "compliance"
        ],
        "summary": "One regime's classification",
        "description": "404 when this model has never been classified under this regime — the `unclassified` state, which appears as a peer value on the model inventory query.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "put": {
        "operationId": "setClassification",
        "tags": [
          "compliance"
        ],
        "summary": "Replace one regime's classification",
        "description": "PUT, not PATCH: the row is replaced whole, so an old basis can never sit underneath a brand-new class. The body carries this regime's enum (euSystemRiskClass/euGpaiTier on `eu_ai_act`, mrmTier on `mrm`) plus the shared fields. Audited as `classification.set` with the regime recorded.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "model": {
        "name": "model",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Error (RFC 9457)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "FactSource": {
        "type": "string",
        "enum": [
          "declared",
          "derived",
          "measured"
        ],
        "description": "How a value was obtained. declared = asserted by the publisher (attributed, not verified); derived = computed by a producer from the model; measured = reported by a runtime or harness, valid within its stated basis."
      },
...
      "Regime": {
        "type": "string",
        "enum": [
          "eu_ai_act",
          "mrm"
        ],
        "description": "One body of rules: the EU AI Act (`eu_ai_act`) or financial model-risk management (`mrm`: SR 26-2, PRA SS1/23, OSFI E-23). Each gets its own row per model, with its own date, author, reason and review cycle, so writing one can never move another's anchor."
      },
      "EUSystemRiskClass": {
        "type": "string",
        "enum": [
          "unclassified",
          "minimal",
          "limited",
          "high_annex_iii",
          "high_annex_i",
          "prohibited"
        ],
        "description": "The risk class of the systems this model is used in, as declared by a person. `unclassified` is a real answer meaning nobody has said yet — it is never rendered as, and never quietly becomes, `minimal`."
      },
      "EUGpaiTier": {
        "type": "string",
        "enum": [
          "none",
          "gpai",
          "gpai_systemic"
        ],
        "description": "The model itself under Art. 53/55. A general-purpose model is not 'high risk' or 'low risk' — that is a different question, answered by euSystemRiskClass."
      },
      "ClassificationState": {
        "type": "string",
        "enum": [
          "unclassified",
          "untiered",
          "unvalidated",
          "stale",
          "current"
        ],
        "description": "Computed per regime on every read, never stored. The `eu_ai_act` ladder is `unclassified`, `stale`, `current`; the `mrm` ladder is `untiered`, `unvalidated`, `stale`, `current` (MRMState). Neither regime's first rung is a kind of `stale`."
      },
      "StaleReason": {
        "type": "string",
        "enum": [
          "review_due_passed",
          "version_published_since",
          "production_changed_since",
          "derivation_since",
          "validation_expired",
          "unmonitored_in_production",
          "conditions_outstanding"
        ],
        "description": "Why a classification went out of date. `production_changed_since` reads updated_at, so editing the production version's description trips it — a deliberate false positive, since for a legal field erring toward 'take another look' is the right direction. For `mrm` rows: `validation_expired`, `version_published_since` (a version published after the validation), `unmonitored_in_production` (in production with no evaluation run since it got there), `conditions_outstanding`."
      },
      "ClassificationInput": {
        "type": "object",
        "description": "A declared assessment. classifiedAt, classifiedBy and source are owned by the server; sending any of them is rejected with 400, following the API-wide rule that unknown fields are refused rather than dropped.",
        "properties": {
          "euSystemRiskClass": {
            "$ref": "#/components/schemas/EUSystemRiskClass"
          },
          "euGpaiTier": {
            "$ref": "#/components/schemas/EUGpaiTier"
          },
          "mrmTier": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MRMTier"
              }
            ],
            "description": "The `mrm` path's enum. Defaults to `untiered`. Sending an EU field on the `mrm` path, or this on the `eu_ai_act` path, is rejected with 400."
          },
          "intendedPurpose": {
            "type": "string",
            "description": "Required on `eu_ai_act` unless euSystemRiskClass is `unclassified` — nobody can review a class with no stated purpose. Optional on `mrm`."
          },
          "basis": {
            "type": "string",
            "description": "Why this class or tier: the reasoning, not the conclusion. Required for high_annex_iii and high_annex_i, and for tier_1 and out_of_scope (422, `details.field: basis`)."
          },
          "reviewDueAt": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Epoch millis, must be in the future. Null means no review scheduled, which is surfaced rather than hidden."
          }
        }
      },
      "Classification": {
        "type": "object",
        "description": "A stored assessment plus the state computed from it at read time.",
        "required": [
          "modelId",
          "regime",
          "classifiedAt",
          "source",
          "state"
        ],
        "properties": {
          "modelId": {
            "type": "string"
          },
          "regime": {
            "$ref": "#/components/schemas/Regime"
          },
          "euGpaiTier": {
            "$ref": "#/components/schemas/EUGpaiTier"
          },
          "euSystemRiskClass": {
            "$ref": "#/components/schemas/EUSystemRiskClass"
          },
          "mrmTier": {
            "$ref": "#/components/schemas/MRMTier"
          },
          "intendedPurpose": {
            "type": "string"
          },
          "basis": {
            "type": "string"
          },
          "classifiedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Server-set. The anchor every staleness check measures against."
          },
          "classifiedBy": {
            "type": "string",
            "description": "Server-set from the actor header at time of write."
          },
          "reviewDueAt": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FactSource"
              }
            ],
            "description": "Always `declared`. There is no computed path for a legal class."
          },
          "state": {
            "$ref": "#/components/schemas/ClassificationState",
            "description": "The row's own regime's ladder: ClassificationState on `eu_ai_act`, MRMState on `mrm`."
          },
          "staleReasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaleReason"
            },
            "description": "Every reason that fired, not the first — someone deciding whether to redo an assessment wants the whole picture."
          },
          "version": {
            "type": "string",
            "description": "`mrm` rows only: the version the state is about — the production version if the model has one, else its newest. Absent when the model has no versions."
          },
          "latestValidation": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Validation"
              }
            ],
            "description": "`mrm` rows only: that version's latest validation, absent when it has none."
          }
        }
      },
      "ClassificationList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Classification"
            }
          }
        },
        "description": "Every regime's row for a model. Not paginated: regimes are a closed, small set."
      },
...
      "MRMTier": {
        "type": "string",
        "enum": [
          "untiered",
          "tier_1",
          "tier_2",
          "tier_3",
          "out_of_scope"
        ],
        "description": "Firm-assigned model-risk tier, `tier_1` highest. `untiered` is the default and is never rendered as low risk. `out_of_scope` is a declared exclusion from the firm's framework (e.g. the SR 26-2 genAI carve-out) and requires a basis."
      },
...
      "ValidationOutcome": {
        "type": "string",
        "enum": [
          "approved",
          "conditional",
          "rejected",
          "undetermined"
        ],
        "description": "`conditional` is fit subject to `conditions`, which are then required. `undetermined` is reviewed but cannot conclude yet."
      },
...
      "Validation": {
        "type": "object",
        "required": [
          "id",
          "versionId",
          "outcome",
          "validatedAt",
          "independenceEvidenced",
          "source"
        ],
        "description": "One independent judgement on one version. Append-only: a re-validation is a new row, and the current answer is the latest.",
        "properties": {
          "id": {
            "type": "string"
          },
          "versionId": {
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/ValidationOutcome"
          },
          "scope": {
            "type": "string"
          },
          "findings": {
            "type": "string"
          },
          "conditions": {
            "type": "string"
          },
          "conditionsClearedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Set once by :clearConditions. While a conditional validation's conditions are uncleared, the state is `stale` with `conditions_outstanding`."
          },
          "validUntil": {
            "type": "integer",
            "format": "int64"
          },
          "evidenceArtifactId": {
            "type": "string"
          },
          "validatedBy": {
            "type": "string",
            "description": "Server-set from the actor header."
          },
          "validatedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Server-set."
          },
          "independenceEvidenced": {
            "type": "boolean",
            "description": "validatedBy is set and differs from the version's author. A flag, never a refusal: a one-person team or a shared service account trips it legitimately, and a human decides whether it is a finding."
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FactSource"
              }
            ],
            "description": "Always `declared`."
          }
        }
      },
...
    }
  }
}